+++ yesod-core/test/YesodCoreTest.hs
@@ -22,6 +22,8 @@ import qualified YesodCoreTest.JsAttributes as JsAttributes
 import qualified YesodCoreTest.JsLoader as JsLoader
 import qualified YesodCoreTest.RequestBodySize as RequestBodySize
 import qualified YesodCoreTest.Json as Json
+import qualified YesodCoreTest.Content as Content
+import qualified YesodCoreTest.NestedDispatch as NestedDispatch
 
 -- Skip on Windows, see https://github.com/yesodweb/yesod/issues/1523#issuecomment-398278450
 #ifndef WINDOWS
@@ -70,3 +72,6 @@ specs = do
       Csrf.csrfSpec
       breadcrumbTest
       metaTest
+      Content.specs
+      describe "NestedDispatch" $ do
+          NestedDispatch.specs